requirements.txt
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
conda==4.3.16
idna==2.8
mysql-connector==2.1.6
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
pep8==1.7.1
pycodestyle==2.5.0
pycosat==0.6.3
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
ruamel.yaml==0.15.91
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.3
textblob==0.15.3
tweepy==3.7.0
urllib3==1.24.2
